Document first crack at digital_ocean module
The [doc](http://docs.ansible.com/ansible/latest/digital_ocean_module.html) implies that this should (re)create the specific Droplet if it does not a) already exist b) in the specified configuration.

We *probably* don't want the `id` parameter. (How do we know if if we're creating it? We care about the name, the uniqueness of the name, and the config.)

But we **never get that far** because we're stopped dead in our tracks by our old friend

```
TASK [Verify that the specified Droplet is running] ********************************************************************************************************************************************************************************************
fatal: [demo2]: FAILED! => {"changed": false, "failed": true, "msg": "dopy >= 0.3.2 required for this module"}
```

When we were doing this entirely on our local system, we knew where the version of `pip` that Ansible was using lived, so we could invoke it directly to add an acceptable version of `dopy`, and off we went. That doesn't work in an [`ansible-silo`](https://github.com/groupon/ansible-silo) environment, and we haven't yet seen anything in the user doc or [development documentation](https://github.com/groupon/ansible-silo#development) in the form of a suggestion on how to proceed. We envision a worst-case scenario of having to fork the `ansible-silo` GitHub repo maintain our own Docker image distinct from [Groupon's](https://hub.docker.com/r/grpn/ansible-silo/).

Fifty steps forward and forty-nine back. Don't forget about the blindfold, and do mind the gap. :tired_face: