Clean up variable definitions (#9)
* Ensure ansible-vault protected files are encrypted via precommit hook

The contents of the file `pre-commit.example` are to be copied to `.git/hooks/pre-commit` and made executable (`chmod 755 .git/hooks/pre-commit`). This will ensure that files whose names match the pattern `.*vault.*\yml$` which *are not* encrypted will cause the commit to abort, with an instructive message.

The Playbooks, and the README file, have been updated appropriately. The README also now contains a brief summary of the convention and the distinction between secret and non-secret files.

Addresses Item 3 of the list in Issue #6.

* Move or remove inappropriate variable definitions from secret var file

This addresses the second item in the list opening Issue #6.

The `droplet_user_name` variable *probably* doesn't need to be secret; it has been moved to the `running_droplet_details` variable file. The `provision_droplet.yml` Playbook was the only Playbook that used the variable without also including the `running_droplet_details` variable file; it now does.

The `ansible_become_pass` variable is no longer used (a hashed password is required instead), so it has been removed.

* Define all non-trivial variables in variable files

This affects the `provision_droplet.yml` Playbook, which gets a bespoke `provisioning_settings` variable file, and `rename_and_retag_droplet.yml`, which sees its `new_droplet_name` variable definition moved to `running_droplet_details`.

[Closes #6]